add policy to every controller

+++ app/Http/Controllers/SpecialWorkingHoursFactorController.php
@@ -5,6 +5,7 @@ namespace App\Http\Controllers;
 use App\Models\Organization;
 use App\Models\SpecialWorkingHoursFactor;
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Gate;
 use Illuminate\Validation\Rule;
 use Inertia\Inertia;
 
@@ -12,11 +13,15 @@ class SpecialWorkingHoursFactorController extends Controller
 {
     public function show(SpecialWorkingHoursFactor $specialWorkingHoursFactor)
     {
+        Gate::authorize('view', SpecialWorkingHoursFactor::class);
+
         return Inertia::render('SWHF/SWHFShow', ['specialWorkingHoursFactor' => $specialWorkingHoursFactor]);
     }
 
     public function store(Request $request)
     {
+        Gate::authorize('create', SpecialWorkingHoursFactor::class);
+
         $validated = $request->validate([
             'type' => ['required', Rule::in(array_keys(SpecialWorkingHoursFactor::$TYPES))],
             'extra_charge' => 'required|numeric',
@@ -37,9 +42,11 @@ class SpecialWorkingHoursFactorController extends Controller
         return back()->with('success', "Besonderer Arbeitszeitzuschlag erfolgreich gespeichert.");
     }
 
-    
+
     public function destroy(Request $request, SpecialWorkingHoursFactor $specialWorkingHoursFactor)
     {
+        Gate::authorize('delete', $specialWorkingHoursFactor);
+
         $specialWorkingHoursFactor->delete();
 
         return back()->with('success', 'Arbeitszuschlag erfolgreich gelöscht.');
